requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.2.1
Faker==37.1.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
whitenoise>=6.7.0
python-decouple>=3.8
